Initial code to parse a module
