+++ src/extract.rs
@@ -1,180 +1,62 @@
-use crate::cleandoc;
-use textwrap;
+use crate::docstrings;
+use rustpython_ast::{Constant, ExprKind, StmtKind};
+use rustpython_parser::parser;
 
-#[derive(Debug)]
-struct Argument {
-    name: String,
-    type_: Option<String>,
-    default: Option<String>,
-    documentation: Option<String>,
+#[derive(Clone)]
+pub struct Function {
+    pub name: String,
+    // TODO: arguments?
+    pub docstring: docstrings::Docstring,
 }
 
-#[derive(Debug)]
-struct Raises {
-    exception: String,
-    documentation: Option<String>,
+pub struct Module {
+    // pub name: String,
+    // pub docstring: docstrings::Docstring,
+    pub functions: Vec<Function>,
 }
 
-struct Docstring {
-    pub title: String,
-    pub description: String,
-    pub returns: String,
-    pub arguments: Vec<Argument>,
-    pub raises: Vec<Raises>,
-}
+pub fn extract(code: &str) -> Module {
+    let python_ast = parser::parse_program(&code, "something").expect("Unable to parse");
 
-fn parse_arguments(docstring: &str) -> Vec<Argument> {
-    let ds = textwrap::dedent(docstring);
+    let mut functions = Vec::new();
 
-    let mut arguments = Vec::new();
-    let mut lines = ds.lines();
-
-    let mut current_argument: Option<Argument> = None;
-
-    // for each line in the docstring, if it has a colon, it's an argument
-    // otherwise, it's part of the argument's documentation
-
-    while let Some(line) = lines.next() {
-        if line.contains(':') {
-            // if we have a current argument, push it to the list
-            if let Some(argument) = current_argument {
-                arguments.push(argument);
-            }
-
-            // start a new argument
-            let mut parts = line.splitn(2, ':');
-            let name = parts.next().unwrap().trim().to_string();
-
-            let documentation = parts.next().map(|s| s.trim().to_string());
-            current_argument = Some(Argument {
+    // find all functions in ast
+    for statement in python_ast {
+        match statement.node {
+            StmtKind::FunctionDef {
                 name,
-                type_: None,
-                default: None,
-                documentation,
-            });
-        } else {
-            // this should not happen, but if it does, just ignore it
-            if current_argument.is_none() {
-                continue;
-            }
+                args,
+                body,
+                decorator_list,
+                returns,
+                type_comment,
+            } => {
+                println!("Found function: {}", name);
 
-            // if we have a current argument, add this line to its documentation
-            let current_argument = current_argument.as_mut().unwrap();
-            let documentation = current_argument
-                .documentation
-                .take()
-                .unwrap_or_else(|| "".to_string());
-            current_argument.documentation = Some((documentation + "\n" + line).trim().to_owned());
+                // find docstring, the first statement in the body of the function
+                // that's an Expr with a Constant value
+
+                let docstring = match body.first() {
+                    Some(stmt) => match &stmt.node {
+                        StmtKind::Expr { value } => match &value.node {
+                            ExprKind::Constant { value, kind: _ } => match value {
+                                Constant::Str(value) => value.clone(),
+                                _ => "".to_string(),
+                            },
+                            _ => "".to_string(),
+                        },
+                        _ => "".to_string(),
+                    },
+                    None => "".to_string(),
+                };
+
+                println!("Docstring: {}", docstring);
+            }
+            _ => {}
         }
     }
 
-    // push the last argument
-    if let Some(argument) = current_argument {
-        arguments.push(argument);
-    }
-
-    arguments
-}
-
-fn parse_raises(docstring: &str) -> Vec<Raises> {
-    let ds = textwrap::dedent(docstring);
-
-    let mut raises = Vec::new();
-    let mut lines = ds.lines();
-
-    let mut current_raises: Option<Raises> = None;
-
-    // for each line in the docstring, if it has a colon, it's an argument
-    // otherwise, it's part of the argument's documentation
-
-    while let Some(line) = lines.next() {
-        if line.contains(':') {
-            // if we have a current argument, push it to the list
-            if let Some(r) = current_raises {
-                raises.push(r);
-            }
-
-            // start a new argument
-            let mut parts = line.splitn(2, ':');
-            let exception = parts.next().unwrap().trim().to_string();
-
-            let documentation = parts.next().map(|s| s.trim().to_string());
-            current_raises = Some(Raises {
-                exception,
-                documentation,
-            });
-        } else {
-            // this should not happen, but if it does, just ignore it
-            if current_raises.is_none() {
-                continue;
-            }
-
-            // if we have a current argument, add this line to its documentation
-            let current_raises = current_raises.as_mut().unwrap();
-            let documentation = current_raises
-                .documentation
-                .take()
-                .unwrap_or_else(|| "".to_string());
-            current_raises.documentation = Some((documentation + "\n" + line).trim().to_owned());
-        }
-    }
-
-    // push the last argument
-    if let Some(r) = current_raises {
-        raises.push(r);
-    }
-
-    raises
-}
-
-impl Docstring {
-    pub fn new_from_string(docstring: &str) -> Self {
-        let cleaned_docstring = cleandoc::cleandoc(docstring);
-
-        let mut description = String::new();
-        let mut arguments = String::new();
-        let mut returns = String::new();
-        let mut raises = String::new();
-
-        let mut current_section = &mut description;
-
-        let mut lines = cleaned_docstring.lines();
-
-        // the title starts with the first line and ends when the first empty line is encountered
-        // pop lines until the first empty line is encountered
-        let title = lines
-            .by_ref()
-            .take_while(|line| !line.trim().is_empty())
-            .collect::<Vec<&str>>()
-            .join(" ");
-
-        for line in lines {
-            if line.starts_with("Args:") {
-                current_section = &mut arguments;
-                continue;
-            } else if line.starts_with("Returns:") {
-                current_section = &mut returns;
-                continue;
-            } else if line.starts_with("Raises:") {
-                current_section = &mut raises;
-                continue;
-            }
-
-            current_section.push_str(line);
-            current_section.push('\n')
-        }
-
-        let arguments = parse_arguments(&arguments);
-        let raises = parse_raises(&raises);
-
-        Self {
-            title,
-            description: description.trim().to_string(),
-            arguments,
-            returns: textwrap::dedent(&returns).trim().to_string(),
-            raises,
-        }
-    }
+    Module { functions }
 }
 
 #[cfg(test)]
@@ -182,82 +64,51 @@ mod tests {
     use pretty_assertions::assert_eq;
 
     #[test]
-    fn it_parses_docstrings() {
-        let docstring = r#"
-        Fetches rows from a Smalltable.
+    fn it_parses_python_function() {
+        let code = r#"
+        def foo(a: int, b: int) -> int:
+            """Example docstring
 
-        Retrieves rows pertaining to the given keys from the Table instance
-        represented by table_handle.  String keys will be UTF-8 encoded.
+            Args:
+                a: a number
+                b: another number
 
-        Args:
-            table_handle: An open smalltable.Table instance.
-            keys: A sequence of strings representing the key of each table
-              row to fetch.  String keys will be UTF-8 encoded.
-            require_all_keys: If True only rows with values set for all keys will be
-              returned.
+            Returns:
+                the sum of a and b
 
-        Returns:
-            A dict mapping keys to the corresponding table row data
-            fetched. Each row is represented as a tuple of strings. For
-            example:
-
-            {b'Serak': ('Rigel VII', 'Preparer'),
-             b'Zim': ('Irk', 'Invader'),
-             b'Lrrr': ('Omicron Persei 8', 'Emperor')}
-
-            Returned keys are always bytes.  If a key from the keys argument is
-            missing from the dictionary, then that row was not found in the
-            table (and require_all_keys must have been False).
-
-        Raises:
-            IOError: An error occurred accessing the smalltable.
+            Raises:
+                ValueError: if a or b are not numbers
+            """
+            return a + b
         "#;
 
-        let parsed_docstring = super::Docstring::new_from_string(docstring);
+        let result = super::extract(&textwrap::dedent(code));
 
-        assert_eq!(parsed_docstring.title, "Fetches rows from a Smalltable.");
+        assert_eq!(result.functions.len(), 1);
+
+        let function = result.functions[0].clone();
+
+        assert_eq!(function.name, "foo");
+        assert_eq!(function.docstring.title, "Example docstring");
+        assert_eq!(function.docstring.description, "");
+        assert_eq!(function.docstring.arguments.len(), 2);
+
+        let arguments = function.docstring.arguments.clone();
+
+        assert_eq!(arguments[0].name, "a");
+        assert_eq!(arguments[0].documentation, Some("a number".to_string()));
+        assert_eq!(arguments[1].name, "b");
         assert_eq!(
-            parsed_docstring.description,
-            "Retrieves rows pertaining to the given keys from the Table instance\nrepresented by table_handle.  String keys will be UTF-8 encoded."
+            arguments[1].documentation,
+            Some("another number".to_string())
         );
 
-        // print parsed_docstring.arguments
-
-        println!("{:#?}", parsed_docstring.arguments);
-
-        assert_eq!(parsed_docstring.arguments.len(), 3);
-        assert_eq!(parsed_docstring.arguments[0].name, "table_handle");
+        assert_eq!(function.docstring.returns, "the sum of a and b");
+        assert_eq!(function.docstring.raises.len(), 1);
+        assert_eq!(function.docstring.raises[0].exception, "ValueError");
         assert_eq!(
-            parsed_docstring.arguments[0].documentation,
-            Some("An open smalltable.Table instance.".to_string())
-        );
-        assert_eq!(parsed_docstring.arguments[0].default, None);
-        assert_eq!(parsed_docstring.arguments[0].type_, None);
-
-        assert_eq!(parsed_docstring.arguments[1].name, "keys");
-        assert_eq!(parsed_docstring.arguments[1].documentation, Some("A sequence of strings representing the key of each table\n  row to fetch.  String keys will be UTF-8 encoded.".to_string()));
-        assert_eq!(parsed_docstring.arguments[1].default, None);
-        assert_eq!(parsed_docstring.arguments[1].type_, None);
-
-        assert_eq!(parsed_docstring.arguments[2].name, "require_all_keys");
-        assert_eq!(
-            parsed_docstring.arguments[2].documentation,
-            Some("If True only rows with values set for all keys will be\n  returned.".to_string())
-        );
-        assert_eq!(parsed_docstring.arguments[2].default, None);
-        assert_eq!(parsed_docstring.arguments[2].type_, None);
-
-        assert_eq!(
-            parsed_docstring.returns,
-            "A dict mapping keys to the corresponding table row data\nfetched. Each row is represented as a tuple of strings. For\nexample:\n\n{b'Serak': ('Rigel VII', 'Preparer'),\n b'Zim': ('Irk', 'Invader'),\n b'Lrrr': ('Omicron Persei 8', 'Emperor')}\n\nReturned keys are always bytes.  If a key from the keys argument is\nmissing from the dictionary, then that row was not found in the\ntable (and require_all_keys must have been False)."
-        );
-
-        assert_eq!(parsed_docstring.raises.len(), 1);
-
-        assert_eq!(parsed_docstring.raises[0].exception, "IOError");
-        assert_eq!(
-            parsed_docstring.raises[0].documentation,
-            Some("An error occurred accessing the smalltable.".to_string())
+            function.docstring.raises[0].documentation,
+            Some("if a or b are not numbers".to_string())
         );
     }
 }
